+++ braintree/dispute_search.py
@@ -4,7 +4,10 @@ class DisputeSearch:
     amount_disputed             =   Search.RangeNodeBuilder("amount_disputed")
     amount_won                  =   Search.RangeNodeBuilder("amount_won")
     case_number                 =   Search.TextNodeBuilder("case_number")
+    # NEXT_MAJOR_VERSION Remove this attribute
+    # DEPRECATED The chargeback_protection_level attribute is deprecated in favor of protection_level
     chargeback_protection_level =   Search.MultipleValueNodeBuilder("chargeback_protection_level")
+    protection_level            =   Search.MultipleValueNodeBuilder("protection_level")
     customer_id                 =   Search.TextNodeBuilder("customer_id")
     disbursement_date           =   Search.RangeNodeBuilder("disbursement_date")
     effective_date              =   Search.RangeNodeBuilder("effective_date")