+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `pre_dispute_program` to `Dispute` and `DisputeSearch`
 * Add `AutoAccepted` status to `Dispute`
 * Add `DisputeAutoAccepted` to `WebhookNotification.Kind`
+* Deprecate `chargeback_protection_level` and add `protection_level` to `Dispute` and `DisputeSearch`
 
 ## 4.17.1
 * Prepare http request before setting url to resolve issue where dot segments get normalized